ubuntu gcc compiler warnings fixed

+++ src/lib/utils/s21_iseven.c
@@ -4,7 +4,8 @@
 #include "s21_math.h"
 
 bool s21_iseven(double x) {
-  if (s21_isint(x) && s21_fabs(s21_fmod(x, 2.0)) < S21_EPSILON) {  // NOLINT
+  if (s21_isint(x) &&
+      s21_fabs((double)s21_fmod(x, 2.0)) < S21_EPSILON) {  // NOLINT
     return true;
   }
   return false;
